Update types to reflect project exports
Love the project & we're using in our own work. Thank you for creating it.

The types do not appear to reflect how the javascript is exported, however. They aren't exported to a default namespace like the types presently represent. This proposed update is what we used so that the type system resolved the exports correctly.